# archive/litgo6_remaining_files/openapi.yaml
openapi: 3.0.3
info:
  title: LITGO5 API
  description: |
    API do Sistema de Matching Jurídico Inteligente LITGO5
    
    ## Visão Geral
    O LITGO5 é uma plataforma que conecta clientes a advogados especializados através de algoritmos de IA e análise semântica.
    
    ## Autenticação
    Todos os endpoints (exceto health check) requerem autenticação JWT via header `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Triagem: 30 requests/minuto
    - Conversação: 60 requests/minuto
    - Status: 120 requests/minuto
    - Outros endpoints: 100 requests/minuto
    
    ## Versioning
    Esta API segue versionamento semântico. Versão atual: v1
    
  version: 1.0.0
  contact:
    name: LITGO5 Team
    email: dev@litgo5.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Desenvolvimento Local
  - url: https://api.litgo5.com
    description: Produção
...
paths:
...
  /api/triage:
    post:
      tags:
        - Triagem
      summary: Triagem Inteligente Assíncrona
      description: Inicia processo de triagem assíncrona de caso com IA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - texto_cliente
                - coords
              properties:
                texto_cliente:
                  type: string
                  example: "Fui demitido sem justa causa e não recebi as verbas rescisórias"
                coords:
                  type: array
                  items:
                    type: number
                  example: [-23.5505, -46.6333]
      responses:
        '202':
          description: Triagem iniciada
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    example: "abc123-def456"
                  status:
                    type: string
                    example: "accepted"
                  message:
                    type: string
                    example: "A triagem do seu caso foi iniciada..."
  /api/triage/status/{task_id}:
    get:
      tags:
        - Triagem
      summary: Status da Triagem
      description: Consulta o status de uma triagem em andamento
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          example: "abc123-def456"
      responses:
        '200':
          description: Status da triagem
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending, processing, completed, failed]
                  result:
                    type: object
                    properties:
                      case_id:
                        type: string
                      area:
                        type: string
                      subarea:
                        type: string
                      urgency_h:
                        type: integer
                      embedding:
                        type: array
                        items:
                          type: number
  /api/match:
    post:
      tags:
        - Matching
      summary: Match de Advogados
      description: Encontra os melhores advogados para um caso específico
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - case_id
              properties:
                case_id:
                  type: string
                  example: "case-789"
                k:
                  type: integer
                  default: 5
                  example: 5
                preset:
                  type: string
                  enum: [fast, balanced, expert]
                  default: balanced
                radius_km:
                  type: number
                  default: 50
                  example: 50
                exclude_ids:
                  type: array
                  items:
                    type: string
                  example: ["lawyer-123"]
      responses:
        '200':
          description: Matches encontrados
          content:
            application/json:
              schema:
                type: object
                properties:
                  case_id:
                    type: string
                  matches:
                    type: array
                    items:
                      $ref: '#/components/schemas/LawyerMatch'
  /api/explain:
    post:
      tags:
        - Matching
      summary: Explicação de Matches
      description: Gera explicações detalhadas sobre os matches
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - case_id
                - lawyer_ids
              properties:
                case_id:
                  type: string
                  example: "case-789"
                lawyer_ids:
                  type: array
                  items:
                    type: string
                  example: ["lw-001", "lw-002"]
      responses:
        '200':
          description: Explicações geradas
          content:
            application/json:
              schema:
                type: object
                properties:
                  explanations:
                    type: object
                    additionalProperties:
                      type: string
                    example:
                      "lw-001": "Dr. João Silva é uma excelente opção! Com 95% de compatibilidade..."
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LawyerMatch:
      type: object
      properties:
        lawyer_id:
          type: string
        nome:
          type: string
        fair:
          type: number
        equity:
          type: number
        features:
          type: object
          properties:
            A:
              type: number
            S:
              type: number
            T:
              type: number
            G:
              type: number
            Q:
              type: number
            U:
              type: number
            R:
              type: number
            C:
              type: number
        breakdown:
          type: object
          properties:
            area_match:
              type: string
            similarity:
              type: string
            geo_score:
              type: string
        weights_used:
          type: object
        preset_used:
          type: string
        avatar_url:
          type: string
        is_available:
          type: boolean
        primary_area:
          type: string
        rating:
          type: number
        distance_km:
          type: number
...